+++ crates/resvg/src/lib.rs
@@ -23,9 +23,87 @@ mod filter;
 mod geom;
 mod image;
 mod mask;
-mod paint_server;
 mod path;
 mod render;
-mod tree;
 
-pub use crate::tree::Tree;
+/// Renders a tree onto the pixmap.
+///
+/// `transform` will be used as a root transform.
+/// Can be used to position SVG inside the `pixmap`.
+///
+/// Text nodes should be already converted into paths using
+/// [`usvg::TreeTextToPath::convert_text`].
+///
+/// The produced content is in the sRGB color space.
+pub fn render(
+    tree: &usvg::Tree,
+    transform: tiny_skia::Transform,
+    pixmap: &mut tiny_skia::PixmapMut,
+) {
+    let target_size = tiny_skia::IntSize::from_wh(pixmap.width(), pixmap.height()).unwrap();
+    let max_bbox = tiny_skia::IntRect::from_xywh(
+        -(target_size.width() as i32) * 2,
+        -(target_size.height() as i32) * 2,
+        target_size.width() * 4,
+        target_size.height() * 4,
+    )
+    .unwrap();
+
+    let ts =
+        usvg::utils::view_box_to_transform(tree.view_box.rect, tree.view_box.aspect, tree.size);
+
+    let root_transform = transform.pre_concat(ts);
+
+    let ctx = render::Context { max_bbox };
+    render::render_nodes(&tree.root, &ctx, root_transform, None, pixmap);
+}
+
+/// Renders a node onto the pixmap.
+///
+/// `transform` will be used as a root transform.
+/// Can be used to position SVG inside the `pixmap`.
+///
+/// Text nodes should be already converted into paths using
+/// [`usvg::TreeTextToPath::convert_text`].
+///
+/// Returns `None` when `node` has a zero size.
+///
+/// The produced content is in the sRGB color space.
+pub fn render_node(
+    node: &usvg::Node,
+    mut transform: tiny_skia::Transform,
+    pixmap: &mut tiny_skia::PixmapMut,
+) -> Option<()> {
+    let bbox = node.abs_bounding_box().and_then(|r| r.to_non_zero_rect())?;
+
+    let target_size = tiny_skia::IntSize::from_wh(pixmap.width(), pixmap.height()).unwrap();
+    let max_bbox = tiny_skia::IntRect::from_xywh(
+        -(target_size.width() as i32) * 2,
+        -(target_size.height() as i32) * 2,
+        target_size.width() * 4,
+        target_size.height() * 4,
+    )
+    .unwrap();
+
+    transform = transform.pre_translate(-bbox.x(), -bbox.y());
+
+    let ctx = render::Context { max_bbox };
+    // TODO: what to do with `text_bbox`?
+    render::render_node(node, &ctx, transform, None, pixmap);
+
+    Some(())
+}
+
+pub(crate) trait OptionLog {
+    fn log_none<F: FnOnce()>(self, f: F) -> Self;
+}
+
+impl<T> OptionLog for Option<T> {
+    #[inline]
+    fn log_none<F: FnOnce()>(self, f: F) -> Self {
+        self.or_else(|| {
+            f();
+            None
+        })
+    }
+}